backend/requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==3.0.5
djangorestframework==3.11.1
djongo==1.3.3
idna==2.10
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
pylint==2.6.0
pymongo==3.11.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.2.4
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
